requirements.txt
chromadb==1.3.0
datasets==4.0.0
einops==0.8.1
langchain==0.3.25
langchain-cohere==0.3.5
langchain-community==0.3.25
langchain-core==0.3.74
langchain-experimental==0.3.4
langchain-huggingface
langchain-ollama==0.3.6
langchain-openai==0.2.14
langchain-text-splitters==0.3.8
libcst==1.8.2
license_expression==30.4.4
llama_index==0.14.7
loguru==0.7.3
matplotlib==3.10.3
ollama==0.6.0
sentence_transformers==5.1.1
torch==2.8.0
tqdm==4.67.1
transformers==4.55.4
tree_sitter==0.25.1
tree_sitter_python==0.23.6
venn==0.1.3
